+++ aether-utils/aether-client/.gitignore
@@ -1,1 +1,2 @@
 dist/
+VERSION
