+++ aether-utils/aether-client/setup.cfg
@@ -2,7 +2,7 @@
 test=pytest
 
 [tool:pytest]
-python_files = aether/client_test/test*.py
+python_files = aether/client/test.py
 addopts = --maxfail=1
 
 [flake8]
@@ -13,6 +13,6 @@ ignore =
 exclude =
     ./.eggs/*
     */__pycache__/*
-    aether/client_test/fixtures/mappings
-    aether/client_test/fixtures/schema
-    aether/client_test/fixtures/submission
+    aether/client/fixtures/mappings
+    aether/client/fixtures/schema
+    aether/client/fixtures/submission
